+++ Videogame/Snake/build.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+pushd "$(dirname "$0")" > /dev/null
+
+mkdir -p bin
+mkdir -p ../../0Common/lib
+
+cd src
+fpc @fpMeta.cfg Snake.pas $@
+ERRCOMP=$?
+
+popd > /dev/null
+
+exit $ERRCOMP